[XENFIN-996] Store ref of collapse element in state.
We want to store the ref to the collapse element in the state, so the component fully rerenders when the collapse element is expanded.
This is necessary so the focus-trap immediately receives the collapse element as well, else it first renders with the collapse element as null.
This affects the items which can be tabbed to. As it only updates the collapse element with the next update, resulting in the user only being table to
tab to the autocomplete list after tabbing through the input box twice.