[XENFIN-1291] Only keep autocomplete backdrop mounted when it is visible
When <Backdrop open={false}>, but it is not unmounted, it still covers
the whole page with its backdrop element that is still mounted.
Unmounting it automatically when it is no longer opened and after its
closing transition has disappeared will ensure that the backdrop element
is not on the page anymore.