[XENFIN-1294] Handle transformation of Date object to internal Picker object
We were passing a raw JS Date object to the DatePicker/DateTimePicker
component. However, the pickers assume that the passed object is one
that is handled by the Picker utils (which is a moment object in
finder-xenit).
In tests, this worked because we were using the lightweight date-fns
library, which uses the javascript Date objects and does not wrap it
with a custom object. Switch it out for the moment objects to catch
these issues in the future when testing.

Test image snapshots changed slightly because date-fns & moment use
different names for day-of-week