[EF2I-1] Do not wipe managed width when we can't find a resize for a column.
When the table is freshly instanciated, it has no resizes that will be sent to this event, even when columns do have widths configured.
We want to avoid wiping out the widths of columns when a resize is triggered, so if we can't find a resize for a column, we keep the original
configuration of the column.